Implementing spliting and signup fetching
