+++ src/infrastructure/services/Api.ts
@@ -1,0 +1,8 @@
+import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
+
+// initialize an empty api service that we'll inject endpoints into later as needed
+export const api = createApi({
+  reducerPath: 'api',
+  baseQuery: fetchBaseQuery({ baseUrl: process.env.REACT_APP_API_URL }),
+  endpoints: () => ({}),
+});
